Adds reserve a car operation
- In carService.js:
First, verifies if the id received is undefined, if so, throws an error.  If the id is valid, it returns the car entity received with the car.reserve value set to "1".

- In carRepository.js:
Prepares the statement for updating the reserved value to "1". Then, executes the statement and returns the modified car entity with getById method.