Adds a basic structure to Reserve Controller, Service and Repository
- In reserveController.js:
Configures the base route by adding a configureRoutes method that (in the future) will contains all route configurations. Additionally, includes an index method that simply renders a basic view with an <h1> tag.

- In reserveService.js:
Defines the ReserveService constructor along with its dependency.

- In reserveRepository.js:
Defines the ReserveRepository constructor along with its dependency.

- In index.html:
Extends the base view to display only an <h1> tag. This is where the list of reserves will be displayed.