Adds validations method in ReserveService
Validation method requires carService as a dependency. This is necessary because the validation for car ID checks if the ID corresponds to a car in the database.

validateDate method verifies if the date is in the format "yyyy-mm-dd" and checks if the received date is within the current year or the next year. Additionally, it validates the days and months.