Updates validateField method and adds getById method in ReserveService
- validateField:
The data received now contains a reserve id instead of a car id. This means that the validation needs to change accordingly. It now verifies if the id is either undefined or an integer. After this update, ReserveService no longer requires a dependency on CarService, so it has been removed.

- getById:
This method verifies if the received id is valid. If it is valid, it returns a Reserve entity using reserveRepository.getById. If the id is not valid, it throws an error.