Updates save method in ReserveRepostiry with editing option
The save method in the Reserve Repository has been updated to accommodate editing functionality.

Now, if it is an update operation, the value of "id" is set to the received reserve id (to return this Reserve entity at the end of the save method). Then, the statement is prepared with the new reserved values. Finally, the statement is executed.