Updates reserve entity and mappers
Modifies reserve entity and mappers to align with ReserveModel. Now, the entity also includes price per day, total price, whether the reserve is paid, and the payment method.

Additionally, updates mappers for the Reserve entity. Renames the function fromDbtoReserveEntity to fromModelToReserveEntity, as this function will now be utilized in ReserveRepository with ReserveModel.