Adds delete method in Reserve Controller, Service and Repository
In userController.js: Implements the deletion of a user with the user id received as a parameter. If successful, notifies the deletion of the user and redirects to "/".

In userService.js: Verifies the validity of the received id. If invalid, throws an error. Otherwise, returns the deleted User entity using the userRepository.delete method.

In userRepository.js: Gets the user to be deleted. If the deletion fails, throws an error. Otherwise, returns the User entity to notify the successful deletion.